changed max_delta from int to float in main_*.py; modified early stopping in callbacks.py; changed early stop criterion from auprc to auroc; created new bash scripts for different downstream tasks

+++ main_linprobe.py
@@ -106,7 +106,7 @@ def get_args_parser():
     # Callback parameters
     parser.add_argument('--patience', default=-1, type=float,
                         help='Early stopping whether val is worse than train for specified nb of epochs (default: -1, i.e. no early stopping)')
-    parser.add_argument('--max_delta', default=0, type=int,
+    parser.add_argument('--max_delta', default=0, type=float,
                         help='Early stopping threshold (val has to be worse than (train+delta)) (default: 0)')
 
     # Criterion parameters
